Decode basyx objs in __deserialize_model()
First we use the basyx decoder as usual to cover all cases where "modelType" is already given in a dict. Then we try to use the specific decoder for this class. We do it for the case when "modelType" is not present in the response data. E.g. get_all_submodel_references()